+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.2.0
+
+- Add useCustomScrollView parameter to pages (allow to use own scroll view, useful for slivers)
+- Rename childBuilder to builder in SubmitPage for consistency with FetchAndSubmitPage
+- Remove useless data param from submit_page app bar builder
+
+## 0.1.2
+
+- Export repository
+
 ## 0.1.1
 
 - Add all real code from private lib